llm_rag_app/requirements.txt
streamlit==1.44.1
langchain==0.3.22
langchain-community==0.3.20
faiss-cpu
python-dotenv
requests
